pyproject.toml
[tool.poetry]
name = "pythonproject2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
generators = "^2020.4.27"
python-dotenv = "^1.0.1"
self = "^2020.12.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.0"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''
[tool.isort]
# максимальная длина строки
line_length = 119

